requirements.txt
amqp==2.3.2
anyjson==0.3.3
billiard==3.5.0.4
boto3==1.7.84
boto==2.49.0
botocore==1.10.84
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
django-celery-results==1.0.1
Django==1.11.13
djangorestframework==3.8.2
docutils==0.14
et-xmlfile==1.0.1
google-api-python-client==1.6.2
googleads==10.1.0
httplib2==0.10.3
idna==2.7
isodate==0.5.4
jdcal==1.3
jmespath==0.9.3
kombu==4.2.1
oauth2client==4.1.2
olefile==0.44
openpyxl==2.4.7
phonenumbers==8.5.2
Pillow==4.2.1
psutil==5.4.3
psycopg2==2.7.1
pyasn1-modules==0.0.10
pyasn1==0.3.1
PySocks==1.6.7
python-dateutil==2.6.0
pytz==2018.4
PyYAML==3.12
requests==2.19.1
rsa==3.4.2
s3transfer==0.1.13
simple-salesforce==0.68.1
six==1.10.0
stripe==1.65.0
suds-jurko==0.6
teamcity-messages==1.17
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
XlsxWriter==0.9.6
xmltodict==0.11.0
